added const

+++ src/main/java/io/taaja/models/generic/Coordinates.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class Coordinates {
-    private int latitude;
-    private int longitude;
-    private int altitude;
+    private float latitude;
+    private float longitude;
+    private Float altitude;
 }
